+++ src/Avalanche/Transaction.cpp
@@ -17,15 +17,15 @@ void BaseTransaction::encode(Data& data) const {
     for (auto byte : BlockchainID) {
         data.push_back(byte);
     }
-    encode32BE(Outputs.size(), data);
+    encode32BE(static_cast<uint32_t>(Outputs.size()), data);
     for (auto output : Outputs) {
         output.encode(data);
     }
-    encode32BE(Inputs.size(), data);
+    encode32BE(static_cast<uint32_t>(Inputs.size()), data);
     for (auto input : Inputs) {
         input.encode(data);
     }
-    encode32BE(Memo.size(), data);
+    encode32BE(static_cast<uint32_t>(Memo.size()), data);
     for (auto byte : Memo) {
         data.push_back(byte);
     }
@@ -36,7 +36,7 @@ void UnsignedCreateAssetTransaction::encode(Data& data) const {
     encodeString(Name, data);
     encodeString(Symbol, data);
     data.push_back(Denomination);
-    encode32BE(InitialStates.size(), data);
+    encode32BE(static_cast<uint32_t>(InitialStates.size()), data);
     for (auto state : InitialStates) {
         state.encode(data);
     }
@@ -44,7 +44,7 @@ void UnsignedCreateAssetTransaction::encode(Data& data) const {
 
 void UnsignedOperationTransaction::encode(Data& data) const {
     BaseTransaction::encode(data);
-    encode32BE(Operations.size(), data);
+    encode32BE(static_cast<uint32_t>(Operations.size()), data);
     for (auto op : Operations) {
         op.encode(data);
     }
@@ -55,7 +55,7 @@ void UnsignedImportTransaction::encode(Data& data) const {
     for (auto byte : SourceChain) {
         data.push_back(byte);
     }
-    encode32BE(ImportInputs.size(), data);
+    encode32BE(static_cast<uint32_t>(ImportInputs.size()), data);
     for (auto in : ImportInputs) {
         in.encode(data);
     }
@@ -66,7 +66,7 @@ void UnsignedExportTransaction::encode(Data& data) const {
     for (auto byte : DestinationChain) {
         data.push_back(byte);
     }
-    encode32BE(ExportOutputs.size(), data);
+    encode32BE(static_cast<uint32_t>(ExportOutputs.size()), data);
     for (auto out : ExportOutputs) {
         out.encode(data);
     }
@@ -76,7 +76,7 @@ void SignedTransaction::encode(Data& data) const {
     encode16BE(CodecID, data);
     UnsignedTransaction.encode(data);
     // do not sort Credentials, they will be paired with inputs by index
-    encode32BE(Credentials.size(), data);
+    encode32BE(static_cast<uint32_t>(Credentials.size()), data);
     for (auto cred : Credentials) {
         cred.encode(data);
     }
